+++ typespec-extension/src/code-model-builder.ts
@@ -146,7 +146,7 @@ import {
 import {
   getClientApiVersions,
   getServiceVersion,
-  operationContainsJsonMergePatch,
+  operationIsJsonMergePatch,
   isPayloadProperty,
   originApiVersion,
   specialHeaderNames,
@@ -154,6 +154,7 @@ import {
   isLroNewPollingStrategy,
   operationIsMultipleContentTypes,
   cloneOperationParameter,
+  operationRefersUnion,
 } from "./operation-utils.js";
 import pkg from "lodash";
 const { isEqual } = pkg;
@@ -169,7 +170,7 @@ export class CodeModelBuilder {
   private codeModel: CodeModel;
 
   readonly schemaCache = new ProcessingCache((type: Type, name: string) => this.processSchemaImpl(type, name));
-  readonly operationCache = new Map<Operation, CodeModelOperation>();
+  readonly typeUnionRefCache = new Map<Type, Union | null | undefined>(); // Union means it ref a Union type, null means it does not ref any Union, nndefined means type visited but not completed
 
   private operationExamples: Map<Operation, any> = new Map<Operation, any>();
 
@@ -515,13 +516,19 @@ export class CodeModelBuilder {
       },
     });
 
-    if (operationContainsJsonMergePatch(op)) {
+    if (operationIsJsonMergePatch(op)) {
       // do not generate convenience method for JSON Merge Patch
-      this.trace(`Operation '${op.operation.name}' contains 'application/merge-patch+json'`);
+      this.trace(`Operation '${op.operation.name}' is 'application/merge-patch+json'`);
     } else if (operationIsMultipleContentTypes(op)) {
       // and multiple content types
       // issue link: https://github.com/Azure/autorest.java/issues/1958#issuecomment-1562558219
       this.trace(`Operation '${op.operation.name}' is multiple content-type`);
+    } else if (
+      operationRefersUnion(this.program, op, this.typeUnionRefCache, (it: Type) => {
+        return getTypeName(it, this.typeNameOptions);
+      })
+    ) {
+      // and Union
     } else {
       const convenienceApiName = this.getConvenienceApiName(operation);
       if (convenienceApiName && !isInternal(this.sdkContext, operation)) {